Move SimCalcs logic into SimulationStep
