+++ src/Metropolis/SimBox.cpp
@@ -4,160 +4,9 @@
 
 #include "SimBox.h"
 #include "GPUCopy.h"
-#include "SimCalcs.h"
 
-void SimBox::keepMoleculeInBox(int molIdx, Real** aCoords, int** molData, int* pIdxes, Real* bSize) {
-  int start = molData[MOL_START][molIdx];
-  int end = start + molData[MOL_LEN][molIdx];
-  int pIdx = pIdxes[molData[MOL_PIDX_START][molIdx]];
 
-  for (int i = 0; i < NUM_DIMENSIONS; i++) {
-    if (aCoords[i][pIdx] < 0) {
-      #pragma acc loop independent
-      for (int j = start; j < end; j++) {
-        aCoords[i][j] += bSize[i];
-      }
-    } else if (aCoords[i][pIdx] > size[i]) {
-      #pragma acc loop independent
-      for (int j = start; j < end; j++) {
-        aCoords[i][j] -= bSize[i];
-      }
-    }
-  }
-}
-
-void SimBox::changeMolecule(int molIdx) {
-  SimCalcs::changeMolecule(molIdx);
-}
-
-void SimBox::rollback(int molIdx) {
-  SimCalcs::rollback(molIdx);
-}
-
-void SimBox::translateAtom(int aIdx, Real dX, Real dY, Real dZ, Real** aCoords) {
-  aCoords[X_COORD][aIdx] += dX;
-  aCoords[Y_COORD][aIdx] += dY;
-  aCoords[Z_COORD][aIdx] += dZ;
-}
-
-void SimBox::rotateAtom(int aIdx, int pivotIdx, Real rotX, Real rotY, Real rotZ, Real** aCoords) {
-  Real pX = aCoords[X_COORD][pivotIdx];
-  Real pY = aCoords[Y_COORD][pivotIdx];
-  Real pZ = aCoords[Z_COORD][pivotIdx];
-
-  translateAtom(aIdx, -pX, -pY, -pZ, aCoords);
-  rotateX(aIdx, rotX, aCoords);
-  rotateY(aIdx, rotY, aCoords);
-  rotateZ(aIdx, rotZ, aCoords);
-  translateAtom(aIdx, pX, pY, pZ, aCoords);
-}
-
-void SimBox::rotateX(int aIdx, Real angleDeg, Real** aCoords) {
-  // Real angleRad = degreesToRadians(angleDeg);
-  Real angleRad = angleDeg * 3.14159265358979 / 180.0;
-  Real oldY = aCoords[Y_COORD][aIdx];
-  Real oldZ = aCoords[Z_COORD][aIdx];
-  aCoords[Y_COORD][aIdx] = oldY * cos(angleRad) + oldZ * sin(angleRad);
-  aCoords[Z_COORD][aIdx] = oldZ * cos(angleRad) - oldY * sin(angleRad);
-}
-
-void SimBox::rotateY(int aIdx, Real angleDeg, Real** aCoords) {
-  // Real angleRad = degreesToRadians(angleDeg);
-  Real angleRad = angleDeg * 3.14159265358979 / 180.0;
-  Real oldZ = aCoords[Z_COORD][aIdx];
-  Real oldX = aCoords[X_COORD][aIdx];
-  aCoords[Z_COORD][aIdx] = oldZ * cos(angleRad) + oldX * sin(angleRad);
-  aCoords[X_COORD][aIdx] = oldX * cos(angleRad) - oldZ * sin(angleRad);
-}
-
-void SimBox::rotateZ(int aIdx, Real angleDeg, Real** aCoords) {
-  // Real angleRad = degreesToRadians(angleDeg);
-  Real angleRad = angleDeg * 3.14159265358979 / 180.0;
-  Real oldX = aCoords[X_COORD][aIdx];
-  Real oldY = aCoords[Y_COORD][aIdx];
-  aCoords[X_COORD][aIdx] = oldX * cos(angleRad) + oldY * sin(angleRad);
-  aCoords[Y_COORD][aIdx] = oldY * cos(angleRad) - oldX * sin(angleRad);
-}
-
-Real SimBox::calcSystemEnergy (Real &subLJ, Real &subCharge) {
-  Real total = subLJ + subCharge;
-
-  for (int mol = 0; mol < numMolecules; mol++) {
-    total += calcMolecularEnergyContribution(subLJ, subCharge, mol, mol);
-  }
-
-  return total;
-}
-
-Real SimBox::calcMolecularEnergyContribution(Real &subLJ, Real &subCharge,
-                                             int currMol, int startMol) {
-  return SimCalcs::calcMolecularEnergyContribution(currMol, startMol);
-}
-
-bool SimBox::moleculesInRange(int p1Start, int p1End, int p2Start, int p2End) {
-  int p1, p2;
-  bool out = false;
-
-  for (int p1Idx = p1Start; p1Idx < p1End; p1Idx++) {
-    p1 = primaryIndexes[p1Idx];
-    for (int p2Idx = p2Start; p2Idx < p2End; p2Idx++) {
-      p2 = primaryIndexes[p2Idx];
-      if (calcAtomDistSquared(p1, p2, atomCoordinates, size) <= cutoff * cutoff) {
-        out = true;
-      }
-    }
-  }
-
-  return out;
-}
-
-Real SimBox::calcMoleculeInteractionEnergy (int m1, int m2) {
-  Real energySum = 0;
-
-  const int m1Start = moleculeData[MOL_START][m1];
-  const int m1End = moleculeData[MOL_LEN][m1] + m1Start;
-
-  const int m2Start = moleculeData[MOL_START][m2];
-  const int m2End = moleculeData[MOL_LEN][m2] + m2Start;
-
-  Real** aData = GPUCopy::atomDataPtr();
-  Real** aCoords = GPUCopy::atomCoordinatesPtr();
-  Real* bSize = GPUCopy::sizePtr();
-
-  #pragma acc parallel loop deviceptr(aData, aCoords, bSize) reduction(+:energySum)
-  for (int i = m1Start; i < m1End; i++) {
-    #pragma acc loop independent
-    for (int j = m2Start; j < m2End; j++) {
-      if (aData[ATOM_SIGMA][i] >= 0 && aData[ATOM_SIGMA][j] >= 0
-          && aData[ATOM_EPSILON][i] >= 0 && aData[ATOM_EPSILON][j] >= 0) {
-
-        const Real r2 = calcAtomDistSquared(i, j, aCoords, bSize);
-        if (r2 == 0.0) {
-          energySum += 0.0;
-        } else {
-
-          const Real sigma = calcBlending(aData[ATOM_SIGMA][i], aData[ATOM_SIGMA][j]);
-          const Real epsilon = calcBlending(aData[ATOM_EPSILON][i], aData[ATOM_EPSILON][j]);
-
-          const Real s2r2 = pow(sigma, 2) / r2;
-          const Real s6r6 = pow(s2r2, 3);
-          const Real s12r12 = pow(s6r6, 2);
-          energySum +=  4.0 * epsilon * (s12r12 - s6r6);
-        }
-
-        if (r2 == 0.0) {
-          energySum += 0.0;
-        } else {
-          const Real e = 332.06;
-          energySum += (aData[ATOM_CHARGE][i] * aData[ATOM_CHARGE][j] * e) / sqrt(r2);
-        }
-
-      }
-    }
-  }
-
-  return (energySum);
-}
+// ----- Experimental -----
 
 Real SimBox::calcAtomDistSquared(int a1, int a2, Real** aCoords, Real* bSize) {
   Real dx = makePeriodic(aCoords[X_COORD][a2] - aCoords[X_COORD][a1], X_COORD, bSize);
